indexOf 가 contains 처럼 쓸 수가 없네;;

+++ src/main/java/fir/sec/thi/doxarpgsystem/Stat.java
@@ -48,11 +48,11 @@ public class Stat {
             List list = new ArrayList();
             String s;
             while ((s = R.readLine()) != null) {
-                list.add(Long.valueOf(Cutter(s)));
+                list.add(Cutter(s));
             }
             R.close();
             for (int count = 0; count < 31; count++) {
-                stat[count] = ((Long) list.get(count)).longValue();
+                stat[count] = (Long) list.get(count);
             }
             return stat;
         } catch (IOException localIoException) {
